CDMInterface

Abstract interface to the CDM, using CDMConnector or DatabaseConnector.

Methods

Public methods

CDMInterface$new()

CDMInterface$validate()

CDMInterface$fetchCohortTable()

CDMInterface$fetchMetadata()

CDMInterface$destroy()

CDMInterface$clone()

Method new()

Initializer method

Usage

CDMInterface$new(
  connectionDetails = NULL,
  cdmSchema = NULL,
  resultSchema = NULL,
  tempEmulationSchema = NULL,
  cdm = NULL
)

Arguments

connectionDetails

(DatabaseConnector::createConnectionDetails(): NULL)
Optional; In congruence with cdmSchema and resultSchema. Ignores cdm.

cdmSchema

(character(1): NULL)
Optional; In congruence with connectionDetails and resultSchema. Ignores cdm.

resultSchema

(character(1): NULL)
Optional; In congruence with connectionDetails and cdmSchema. Ignores cdm.

tempEmulationSchema

Schema used to emulate temp tables.

cdm

(CDMConnector::cdm_from_con(): NULL)
Optional; Ignores connectionDetails, cdmSchema, and resultSchema.

Returns

(invisible(self))

Method validate()

Validation method

Usage

CDMInterface$validate()

Returns

(invisible(self))

Method fetchCohortTable()

Fetch specified cohort IDs from a specified cohort table

Usage

CDMInterface$fetchCohortTable(
  cohorts,
  cohortTableName,
  andromeda,
  andromedaTableName,
  minEraDuration = NULL
)

Arguments

cohorts

(data.frame())
Data frame containing the following columns and data types:

cohortId numeric(1)

Cohort ID's of the cohorts to be used in the cohort table.

cohortName character(1)

Cohort names of the cohorts to be used in the cohort table.

type character(1) ["target", "event', "exit"]

Cohort type, describing if the cohort is a target, event, or exit cohort

cohortTableName

(character(1))
Cohort table name.

andromeda

(Andromeda::andromeda()) Andromeda object.

andromeda

(Andromeda::andromeda()) Andromeda object.

andromedaTableName

(character(1))
Name of the table in the Andromeda object where the data will be loaded.

minEraDuration

(integer(1): 0)
Minimum time an event era should last to be included in analysis

Returns

(data.frame)

Method fetchMetadata()

Fetch metadata from CDM

Usage

CDMInterface$fetchMetadata(andromeda)

Arguments

andromeda

(Andromeda::andromeda()) Andromeda object.

andromeda

(Andromeda::andromeda()) Andromeda object.

Returns

(invisible(NULL))

Method destroy()

Destroys instance

Usage

CDMInterface$destroy()

Returns

(NULL)

Method clone()

The objects of this class are cloneable with this method.

Usage

CDMInterface$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.