InputHandler

Class to handle input from the user. Supports direct paths or input fields through setDataPath().

Super class

TreatmentPatterns::Module -> InputHandler

Active bindings

reactiveValues

(reactiveValues)
reactiveValues class created by reactiveValues.

Methods

Public methods

InputHandler$uiMenu()

InputHandler$uiBody()

InputHandler$server()

InputHandler$uiDatabaseSelector()

InputHandler$setDataPath()

InputHandler$clone()

Inherited methods

TreatmentPatterns::Module$initialize()

TreatmentPatterns::Module$validate()

Method uiMenu()

Method to include a menuItem to link to the body.

Usage

InputHandler$uiMenu(label = "File upload", tag = "fileUpload")

Arguments

label

(character(1))
Label to show for the menuItem.

tag

(character(1))
Tag to use internally in input.

Returns

(menuItem)

Method uiBody()

Method to include a tabItem to include the body.

Usage

InputHandler$uiBody()

Returns

(tabItem)

Method server()

Method to handle the back-end.

Usage

InputHandler$server(input, output, session)

Arguments

input

(input)
Input from the server function.

output

(output)
Output from the server function.

session

(session)
Session from the server function.

Returns

(NULL)

Method uiDatabaseSelector()

Method to include a uiOutput to select between multiple uploaded files.

Usage

InputHandler$uiDatabaseSelector()

Returns

(uiOutput)

Method setDataPath()

Method to dictate where the data is coming from, either from the input through the shiny application, or from a specified path. When one is provided, the other is ignored.

Usage

InputHandler$setDataPath(tag = "uploadField", input = NULL, path = NULL)

Arguments

tag

(character(1))
Tag to use internally in input.

input

(input)
Input from the server function of the shiny app.

path

(character(1))
Path to a zip-file containing TreatmentPatterns output files.

Returns

(invisible(self))

Method clone()

The objects of this class are cloneable with this method.

Usage

InputHandler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.