TreatmentPatterns: Analyzes Real-World Treatment Patterns of a Study Population of Interest

Computes treatment patterns within a given cohort using the Observational Medical Outcomes Partnership (OMOP) common data model (CDM). As described in Markus, Verhamme, Kors, and Rijnbeek (2022) doi:10.1016/j.cmpb.2022.107081 .

See also

Author

Maintainer: Maarten van Kessel m.l.vankessel@erasmusmc.nl (ORCID)

Authors:

Aniek Markus (ORCID)