export

Export andromeda generated by computePathways object to sharable csv-files and/or a zip archive.

export(
  andromeda,
  outputPath,
  ageWindow = 10,
  minCellCount = 5,
  censorType = "minCellCount",
  archiveName = NULL
)

Arguments

andromeda

(Andromeda::andromeda()) Andromeda object.

outputPath

(character(1))

ageWindow

(integer(n): 10)
Number of years to bin age groups into. It may also be a vector of integers. I.e. c(0, 18, 150) which will results in age group 0-18 which includes subjects < 19. And age group 18-150 which includes subjects > 18.

minCellCount

(integer(1): 5)
Minimum count required per pathway. Censors data below x as <x. This minimum value will carry over to the sankey diagram and sunburst plot.

censorType

(character(1))

"minCellCount"

Censors pathways <minCellCount to minCellCount.

"remove"

Censors pathways <minCellCount by removing them completely.

"mean"

Censors pathways <minCellCount to the mean of all frequencies below minCellCount

archiveName

(character(1): NULL)
If not NULL adds the exported files to a ZIP-file with the specified archive name.

Value

(invisible(NULL))

Examples

# \donttest{
library(TreatmentPatterns)
library(CDMConnector)
library(dplyr)

if (require("CirceR", character.only = TRUE, quietly = TRUE)) {
  withr::local_envvar(
    EUNOMIA_DATA_FOLDER = Sys.getenv("EUNOMIA_DATA_FOLDER", unset = tempfile())
  )

  downloadEunomiaData(overwrite = TRUE)

  con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())
  cdm <- cdmFromCon(con, cdmSchema = "main", writeSchema = "main")

  cohortSet <- readCohortSet(
    path = system.file(package = "TreatmentPatterns", "exampleCohorts")
  )

  cdm <- generateCohortSet(
    cdm = cdm,
    cohortSet = cohortSet,
    name = "cohort_table"
  )

  cohorts <- cohortSet %>%
    # Remove 'cohort' and 'json' columns
    select(-"cohort", -"json") %>%
    mutate(type = c("event", "event", "event", "event", "exit", "event", "event", "target")) %>%
    rename(
      cohortId = "cohort_definition_id",
      cohortName = "cohort_name",
    ) %>%
    select("cohortId", "cohortName", "type")

  outputEnv <- computePathways(
    cohorts = cohorts,
    cohortTableName = "cohort_table",
    cdm = cdm
  )

  export(
    andromeda = outputEnv,
    outputPath = tempdir()
  )

  Andromeda::close(outputEnv)
  DBI::dbDisconnect(con, shutdown = TRUE)
}
#> 
#> Download completed!
#> ℹ Generating 8 cohorts
#> ℹ Generating cohort (1/8) - Acetaminophen)
#> ✔ Generating cohort (1/8) - Acetaminophen) [126ms]
#> 
#> ℹ Generating cohort (2/8) - Amoxicillin)
#> ✔ Generating cohort (2/8) - Amoxicillin) [97ms]
#> 
#> ℹ Generating cohort (3/8) - Aspirin)
#> ✔ Generating cohort (3/8) - Aspirin) [95ms]
#> 
#> ℹ Generating cohort (4/8) - Clavulanate)
#> ✔ Generating cohort (4/8) - Clavulanate) [96ms]
#> 
#> ℹ Generating cohort (5/8) - Death)
#> ✔ Generating cohort (5/8) - Death) [48ms]
#> 
#> ℹ Generating cohort (6/8) - Doxylamine)
#> ✔ Generating cohort (6/8) - Doxylamine) [80ms]
#> 
#> ℹ Generating cohort (7/8) - PenicillinV)
#> ✔ Generating cohort (7/8) - PenicillinV) [107ms]
#> 
#> ℹ Generating cohort (8/8) - ViralSinusitis)
#> ✔ Generating cohort (8/8) - ViralSinusitis) [128ms]
#> 
#> 
#> Construct treatment pathways, this may take a while for larger datasets.
#> Original number of rows: 8352
#> After eraCollapseSize: 0
#> Selected 1544 
#> out of 8352 rows
#> Iteration: 1
#> Switches: 8352
#> FRFS Combinations: 4
#> LRFS Combinations: 1527
#> Selected 139 
#> out of 6816 rows
#> Iteration: 2
#> Switches: 6816
#> FRFS Combinations: 1
#> LRFS Combinations: 135
#> Selected 1 
#> out of 6682 rows
#> Iteration: 3
#> Switches: 6682
#> FRFS Combinations: 1
#> LRFS Combinations: 0
#> After combinationWindow: 6682
#> Time needed to execute combination window 0.0927278677622477
#> Order the combinations.
#> After filterTreatments: 6679
#> Adding drug sequence number.
#> After maxPathLength: 6675
#> Adding concept names.
#> Ordering the combinations.
#> constructPathways done.
#> Writing metadata to C:\Users\MVANKE~1\AppData\Local\Temp\Rtmp0CfUf4/metadata.csv
#> Writing treatmentPathways to C:\Users\MVANKE~1\AppData\Local\Temp\Rtmp0CfUf4/treatmentPathways.csv
#> Censoring 11840 pathways with a frequency <5 to 5.
#> Writing summaryStatsTherapyDuration to C:\Users\MVANKE~1\AppData\Local\Temp\Rtmp0CfUf4/summaryStatsTherapyDuration.csv
#> Writing countsYearPath to C:\Users\MVANKE~1\AppData\Local\Temp\Rtmp0CfUf4/countsYear.csv
#> Writing countsAgePath to C:\Users\MVANKE~1\AppData\Local\Temp\Rtmp0CfUf4/countsAge.csv
#> Writing countsSexPath to C:\Users\MVANKE~1\AppData\Local\Temp\Rtmp0CfUf4/countsSex.csv
# }